backend/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
black==24.8.0
click==8.1.7
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.1.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
zipp==3.20.1
